chore(shared): extract shared types and validations into packages/shared
