+++ apps/mobile/package.json
@@ -29,7 +29,8 @@
     "react-native-web": "^0.21.0",
     "tailwind-merge": "^3.4.0",
     "tailwindcss": "^3.4.17",
-    "zod": "^4.3.4"
+    "zod": "^4.3.4",
+    "@resave/shared": "workspace:*"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
