+++ apps/mobile/tsconfig.json
@@ -6,6 +6,15 @@
     "paths": {
       "@/*": [
         "./*"
+      ],
+      "@resave/shared": [
+        "../../packages/shared/src/index.ts"
+      ],
+      "@resave/shared/types": [
+        "../../packages/shared/src/types/index.ts"
+      ],
+      "@resave/shared/validations": [
+        "../../packages/shared/src/validations/index.ts"
       ]
     }
   },
