+++ apps/mobile/validations/card.ts
