+++ packages/shared/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@resave/shared",
+  "version": "0.0.1",
+  "private": true,
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./types": "./src/types/index.ts",
+    "./validations": "./src/validations/index.ts"
+  },
+  "dependencies": {
+    "zod": "^4.3.4"
+  }
+}
