+++ packages/shared/src/validations/user.ts
@@ -1,0 +1,60 @@
+import { z } from 'zod';
+
+export const userSchema = z.object({
+  id: z.string().uuid(),
+  email: z.string().email('有効なメールアドレスを入力してください'),
+  createdAt: z.string().datetime(),
+});
+
+export const loginSchema = z.object({
+  email: z
+    .string()
+    .min(1, 'メールアドレスを入力してください')
+    .email('有効なメールアドレスを入力してください'),
+  password: z.string().min(1, 'パスワードを入力してください'),
+});
+
+const passwordSchema = z
+  .string()
+  .min(8, 'パスワードは8文字以上で入力してください')
+  .regex(
+    /^(?=.*[a-zA-Z])(?=.*[0-9])/,
+    'パスワードは英字と数字を含める必要があります'
+  );
+
+export const signupSchema = z
+  .object({
+    email: z
+      .string()
+      .min(1, 'メールアドレスを入力してください')
+      .email('有効なメールアドレスを入力してください'),
+    password: passwordSchema,
+    confirmPassword: z.string().min(1, 'パスワード（確認）を入力してください'),
+  })
+  .refine((data) => data.password === data.confirmPassword, {
+    message: 'パスワードが一致しません',
+    path: ['confirmPassword'],
+  });
+
+export const resetPasswordSchema = z.object({
+  email: z
+    .string()
+    .min(1, 'メールアドレスを入力してください')
+    .email('有効なメールアドレスを入力してください'),
+});
+
+export const updatePasswordSchema = z
+  .object({
+    password: passwordSchema,
+    confirmPassword: z.string().min(1, 'パスワード（確認）を入力してください'),
+  })
+  .refine((data) => data.password === data.confirmPassword, {
+    message: 'パスワードが一致しません',
+    path: ['confirmPassword'],
+  });
+
+export type User = z.infer<typeof userSchema>;
+export type LoginInput = z.infer<typeof loginSchema>;
+export type SignupInput = z.infer<typeof signupSchema>;
+export type ResetPasswordInput = z.infer<typeof resetPasswordSchema>;
+export type UpdatePasswordInput = z.infer<typeof updatePasswordSchema>;
